2008-05-14 20:20 UTC+0700 Andi Jahja <harbour/AT/cbn/net/id>   * makefile.gc     + compiler flag -fno-strict-aliasing to fix compiling       when a pointer is passed-by-reference
  * makefile.bc
    + /C flag to ARFLAGS (don't ignore symbol case)

  * source/common/hbstr.c
  * contrib/libnf/getenvrn.c
  * contrib/hbzlib/zip.c
  * contrib/hbzlib/zipnew.cpp
  * contrib/hbzlib/zipfile.cpp
  * contrib/hbzlib/zipstorage.cpp
  * contrib/wvtgui/wvtcore.c
  * contrib/what32/source/_winsock.c
    ! fixed warnings under GCC 4.3

  * source/vm/hvm.c
    ! casting on initialization of hb_vm_BreakBlock (NULL->0)

  * source/compiler/hbusage.c
    ! do not harcode HB_VER_CVSID, use function instead

+++ xharbour/source/common/hbstr.c
@@ -1,5 +1,5 @@
 /*
- * $Id: hbstr.c,v 1.27 2007/08/28 19:04:55 ronpinkas Exp $
+ * $Id: hbstr.c,v 1.28 2007/12/22 19:04:32 likewolf Exp $
  */
 
 /*
@@ -375,7 +375,7 @@ static double hb_numPow10( int nPrecision )
       }
       else if ( nPrecision > -16 )
       {
-         return 1.0 / s_dPow10[ -nPrecision ];
+         return 1.0 / s_dPow10[ (unsigned int) -nPrecision ];
       }
    }
